Extract need_to_change_pw and sync_possible from __init__
